requirements.txt
PyYAML==5.4.1
numpy==1.24.4
open3d==0.18.0
opencv-python==4.10.0.84
scikit-learn==0.24.1
scipy==1.10.1
segmentation-models-pytorch==0.3.1
tensorboard==2.10.0
tqdm==4.67.1